+++ Sources/Filters.swift
@@ -7,11 +7,15 @@
 import Foundation
 import Stencil
 
-enum FilterError: Error {
-  case invalidInputType
-}
+// For retro-compatibility. Remove in next major.
+@available(*, deprecated: 1.0.2, renamed: "Filters.Error", message: "Use the Filters.Error nested type instead")
+typealias FilterError = Filters.Error
 
 enum Filters {
+  enum Error: Swift.Error {
+    case invalidInputType
+  }
+
   /// Parses filter arguments for a boolean value, where true can by any one of: "true", "yes", "1", and
   /// false can be any one of: "false", "no", "0". If optional is true it means that the argument on the filter is
   /// optional and it's not an error condition if the argument is missing or not the right type
@@ -23,7 +27,7 @@ enum Filters {
   static func parseBool(from arguments: [Any?], index: Int, required: Bool = true) throws -> Bool? {
     guard index < arguments.count, let boolArg = arguments[index] as? String else {
       if required {
-        throw FilterError.invalidInputType
+        throw Error.invalidInputType
       } else {
         return nil
       }
@@ -35,16 +39,16 @@ enum Filters {
     case "true", "yes", "1":
       return true
     default:
-      throw FilterError.invalidInputType
+      throw Error.invalidInputType
     }
   }
 }
 
 struct ArrayFilters {
   static func join(_ value: Any?) throws -> Any? {
-    guard let array = value as? [Any] else { throw FilterError.invalidInputType }
+    guard let array = value as? [Any] else { throw Filters.Error.invalidInputType }
     let strings = array.flatMap { $0 as? String }
-    guard array.count == strings.count else { throw FilterError.invalidInputType }
+    guard array.count == strings.count else { throw Filters.Error.invalidInputType }
 
     return strings.joined(separator: ", ")
   }