+++ Sources/NumFilters.swift
@@ -9,17 +9,17 @@ import Stencil
 
 struct NumFilters {
   static func hexToInt(_ value: Any?) throws -> Any? {
-    guard let value = value as? String else { throw FilterError.invalidInputType }
+    guard let value = value as? String else { throw Filters.Error.invalidInputType }
     return Int(value, radix:  16)
   }
 
   static func int255toFloat(_ value: Any?) throws -> Any? {
-    guard let value = value as? Int else { throw FilterError.invalidInputType }
+    guard let value = value as? Int else { throw Filters.Error.invalidInputType }
     return Float(value) / Float(255.0)
   }
 
   static func percent(_ value: Any?) throws -> Any? {
-    guard let value = value as? Float else { throw FilterError.invalidInputType }
+    guard let value = value as? Float else { throw Filters.Error.invalidInputType }
 
     let percent = Int(value * 100.0)
     return "\(percent)%"